MultiSelectItem is moved under MultiSelect directory, checkedItems is changed to Map
