+++ src/index.js
@@ -12,4 +12,4 @@ export * from './layouts/application/index.js';
 export * from './extended-search/index.js';
 export * from './wizard/index.js';
 export * from './multi-select/index.js';
-export * from './multi-select-item/index.js';
+export * from './multi-select/multi-select-item/index.js';
