+++ src/multi-select/multi-select-item/index.js
