+++ src/multi-select/multi-select-item/multi-select-item.component.scss
