--- src/multi-select/multi-select.component.scss
@@ -1,3 +1,0 @@
-.oc-multi-select {
-  height: 100%;
-}
